@@ -1,5 +1,6 @@
 package com.studyhub.sth.dtos.conteudoEstudo;
 
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -8,7 +9,7 @@ import java.util.UUID;
 @Getter
 @Setter
 public class ConteudoEstudoDto {
-    private UUID id;
+    private UUID conteudoEstudoId;
     private String link;
     private UUID roomId;
 }
